// frontend/src/core/api/generated/model/movementRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */
import type { MovementRequestPurpose } from './movementRequestPurpose'
import type { MovementRequestType } from './movementRequestType'

export interface MovementRequest {
  id: string
  type: MovementRequestType
  /** @nullable */
  materialId?: string | null
  /** @exclusiveMinimum 0 */
  quantity: number
  /** @nullable */
  sourceId?: string | null
  /** @nullable */
  destinationId?: string | null
  /** @nullable */
  partyId?: string | null
  /**
   * @minLength 0
   * @maxLength 120
   * @nullable
   */
  partyName?: string | null
  /** @nullable */
  purpose?: MovementRequestPurpose
  /** @nullable */
  returnOfId?: string | null
  day: string
  /** @nullable */
  expectedReturnDate?: string | null
  /**
   * @minLength 0
   * @maxLength 300
   * @nullable
   */
  returnNote?: string | null
  /**
   * @minLength 0
   * @maxLength 120
   * @nullable
   */
  usageArea?: string | null
  /**
   * @minLength 0
   * @maxLength 500
   * @nullable
   */
  description?: string | null
  /** @nullable */
  inTransit?: boolean | null
  /** @nullable */
  pendingCheck?: boolean | null
  /** @nullable */
  reflectToField?: boolean | null
}